Move findUnbanInformationForMember to UnbanUsers.tsx
